Perform the fine-grained authorization

+++ components/org.wso2.carbon.identity.scim2.common/src/main/java/org/wso2/carbon/identity/scim2/common/impl/SCIMRoleManager.java
@@ -95,7 +95,19 @@ public class SCIMRoleManager implements RoleManager {
     }
 
     @Override
-    public Role createRole(Role role) throws CharonException, ConflictException, BadRequestException {
+    public Role createRole(Role role) throws CharonException, ConflictException,
+            BadRequestException, ForbiddenException {
+
+        List<String> authorizedScopes = (List<String>) IdentityUtil.threadLocalProperties.get().get(
+                SCIMCommonConstants.AUTHORIZED_SCOPES);
+
+        if (authorizedScopes == null ||
+                !(authorizedScopes.contains("internal_role_mgt_create") ||
+                        authorizedScopes.contains("internal_bulk_resource_create") ||
+                        authorizedScopes.contains("internal_bulk_role_create"))) {
+            throw new ForbiddenException("Operation is not permitted. You do not have permissions to" +
+                    " make this request..");
+        }
 
         if (log.isDebugEnabled()) {
             log.debug("Creating role: " + role.getDisplayName());
@@ -188,7 +200,19 @@ public class SCIMRoleManager implements RoleManager {
     }
 
     @Override
-    public void deleteRole(String roleID) throws CharonException, NotFoundException, BadRequestException {
+    public void deleteRole(String roleID) throws CharonException, NotFoundException,
+            BadRequestException, ForbiddenException{
+
+        List<String> authorizedScopes = (List<String>) IdentityUtil.threadLocalProperties.get().get(
+                SCIMCommonConstants.AUTHORIZED_SCOPES);
+
+        if (authorizedScopes == null ||
+                !(authorizedScopes.contains("internal_role_mgt_delete") ||
+                        authorizedScopes.contains("internal_bulk_resource_create") ||
+                        authorizedScopes.contains("internal_bulk_role_delete"))) {
+            throw new ForbiddenException("Operation is not permitted. You do not have permissions to" +
+                    " make this request..");
+        }
 
         try {
             roleManagementService.deleteRole(roleID, tenantDomain);
@@ -373,7 +397,18 @@ public class SCIMRoleManager implements RoleManager {
 
     @Override
     public Role updateRole(Role oldRole, Role newRole)
-            throws BadRequestException, CharonException, ConflictException, NotFoundException {
+            throws BadRequestException, CharonException, ConflictException, NotFoundException, ForbiddenException {
+
+        List<String> authorizedScopes = (List<String>) IdentityUtil.threadLocalProperties.get().get(
+                SCIMCommonConstants.AUTHORIZED_SCOPES);
+
+        if (authorizedScopes == null ||
+                !(authorizedScopes.contains("internal_role_mgt_update") ||
+                        authorizedScopes.contains("internal_bulk_resource_create") ||
+                        authorizedScopes.contains("internal_bulk_role_update"))) {
+            throw new ForbiddenException("Operation is not permitted. You do not have permissions to" +
+                    " make this request..");
+        }
 
         doUpdateRoleName(oldRole, newRole);
         doUpdateUsers(oldRole, newRole);
@@ -554,6 +589,17 @@ public class SCIMRoleManager implements RoleManager {
     @Override
     public Role patchRole(String roleId, Map<String, List<PatchOperation>> patchOperations)
             throws BadRequestException, CharonException, ConflictException, NotFoundException, ForbiddenException {
+
+        List<String> authorizedScopes = (List<String>) IdentityUtil.threadLocalProperties.get().get(
+                SCIMCommonConstants.AUTHORIZED_SCOPES);
+
+        if (authorizedScopes == null ||
+                !(authorizedScopes.contains("internal_role_mgt_update") ||
+                        authorizedScopes.contains("internal_bulk_resource_create") ||
+                        authorizedScopes.contains("internal_bulk_role_update"))) {
+            throw new ForbiddenException("Operation is not permitted. You do not have permissions to" +
+                    " make this request..");
+        }
 
         String currentRoleName = getCurrentRoleName(roleId, tenantDomain);
